==> 4-HSCR_differential_expression_analysis.Rmd <==
---
title: "HSCR Differential Expression Analysis"
author: "Liz Vincent"
date: "`r format(Sys.time(), '%B %d, %Y')`"
output: pdf_document
---
```{r include=FALSE}
knitr::opts_chunk$set(message=FALSE,warning=FALSE,error=FALSE,fig.align="center")
source("init.R")
```

# Differential gene expression testing
## y ~ genotype
```{r genotype}
genotype_diff_test <- differentialGeneTest(dat.filtered[expressed_genes,], fullModelFormulaStr = "~genotype + batch + total_mRNAs", reducedModelFormulaStr = "~batch + total_mRNAs", verbose = T)

genotype_diff_test_sorted <- genotype_diff_test[order(genotype_diff_test$qval),]

write.csv(genotype_diff_test_sorted[genotype_diff_test_sorted$qval < 0.05, c("gene_short_name", "gene_id", "pval", "qval")], file = "DEG_all_cells_genotype.csv", quote = F, row.names = F)

genotype_diff_test_sig_genes <- rownames(genotype_diff_test_sorted[genotype_diff_test_sorted$qval < 0.05,])

heatmap_data <- log10(exprs(dat.filtered)[genotype_diff_test_sig_genes,] + 1) # log transform
heatmap_data <- t(apply(heatmap_data, 1, function(x){ # Normalize gene expr by dividing by max
  x/max(x)
}))

rownames(heatmap_data) <- lookupGeneName(dat.filtered, rownames(heatmap_data))

heatmap_data <- heatmap_data[order.dendrogram(as.dendrogram(hclust(dist(heatmap_data)))),order.dendrogram(as.dendrogram(hclust(dist(t(heatmap_data)))))]

#heatmap_data[heatmap_data == 0] <- NA

#myLower <- 0
#myUpper <- 1

#heatmap_data[heatmap_data < myLower] <- myLower
#heatmap_data[heatmap_data > myUpper] <- myUpper

heatmap_annotation <- pData(dat.filtered)[, "genotype", drop = F]
colnames(heatmap_annotation) <- "Genotype"

heatmap_colors <- list(
  "Genotype" = c("het" = "#E41A1C", "hom" = "#377EB8"))

pdf("DEG_all_cells_genotype_heatmap.pdf", height = 20, width = 20)
pheatmap(mat = heatmap_data,
         scale="none",
         show_rownames = TRUE,
         show_colnames = FALSE,
         drop_levels = FALSE,
         cluster_rows = F,
         cluster_cols = F,
         color = viridis(100),
         #breaks=seq(0, myUpper, length = 100),
         #breaks = seq(0, 2.8, length = 200),
         annotation_col = heatmap_annotation,
         annotation_colors = heatmap_colors,
         annotation_names_col = TRUE,
         annotation_legend = TRUE,
         fontsize_row = 5
#         treeheight_col = 0,
#         treeheight_row = 0
         #cutree_rows = 18,
         #cutree_cols = 9
         )
dev.off()

egENSEMBL <- toTable(org.Mm.egENSEMBL)

genotype_diff_test_entrez <- gsub("\\..*","", genotype_diff_test_sig_genes)
m <- match(genotype_diff_test_entrez, egENSEMBL$ensembl_id)
genotype_diff_test_entrez <- egENSEMBL$gene_id[m]
ego_genotype <- enrichGO(gene = genotype_diff_test_entrez,
                OrgDb =       'org.Mm.eg.db',
                ont           = "BP",
                pAdjustMethod = "BH",
                pvalueCutoff  = 1,
                qvalueCutoff  = 0.05, 
                readable      = TRUE)

write.csv(ego_genotype@result[ego_genotype@result$qvalue < 0.05,], "DEG_all_cells_genotype_GO.csv")

ego_genotype_subset <- ego_genotype@result[ego_genotype@result$ID %in% c("GO:0007219", "GO:0042063", "GO:0045165", "GO:0061564", "GO:0050768", "GO:0045666", "GO:0001505", "GO:0006584", "GO:0061351", "GO:0014033", "GO:0045787", "GO:0030335", "GO:0070371", "GO:0042552", "GO:0009952", "GO:0010718", "GO:0098727", "GO:0043405", "GO:0051302", "GO:0050808", "GO:1990138", "GO:0007088"),]

ego_genotype_subset$Description <- factor(ego_genotype_subset$Description, levels = rev(ego_genotype_subset$Description))

pdf("DEG_all_cells_genotype_GO_plot.pdf", height = 4, width = 8, useDingbats = F)
ggplot(ego_genotype_subset) +
  geom_col(aes(x = Description, y = Count, fill = qvalue)) +
  scale_x_discrete("GO:BP Term") +
  scale_y_continuous("# Genes", expand = c(0,0)) +
  scale_fill_viridis("q-value", limits = c(0, 0.05), breaks = c(0, 0.05), option = "plasma") +
  coord_flip() +
  theme(legend.position = c(0.7, 0.2), 
        legend.background = element_blank(),
        legend.title = element_text(size = 12),
        legend.text = element_text(size = 10),
        axis.text = element_text(size = 10), 
        axis.title = element_text(size = 12),
        axis.title.y = element_text(margin = margin(r = -10)),
        axis.title.x = element_text(margin = margin(t = -2)))
dev.off()
```

```{r genotype:sex}
genotype_sex_diff_test <- differentialGeneTest(dat.filtered[expressed_genes,], fullModelFormulaStr = "~genotype*sex + batch + total_mRNAs", reducedModelFormulaStr = "~genotype + sex + batch + total_mRNAs", verbose = T)

genotype_sex_diff_test_sorted <- genotype_sex_diff_test[order(genotype_sex_diff_test$qval),]

write.csv(genotype_sex_diff_test_sorted[genotype_sex_diff_test_sorted$qval < 0.05, c("gene_short_name", "gene_id", "pval", "qval")], file = "DEG_all_cells_genotypeXsex.csv", quote = F, row.names = F)

genotype_sex_diff_test_sig_genes <- rownames(genotype_sex_diff_test_sorted[genotype_sex_diff_test_sorted$qval < 0.05,])

length(genotype_sex_diff_test_sig_genes)

#tmp <- t(scale(t(exprs(dat.filtered[genotype_sex_diff_test_sorted_sig_genes,])),scale=T,center=T))
#heatmap_data <- log(tmp -min(tmp) + 1)

heatmap_data <- log10(exprs(dat.filtered)[genotype_sex_diff_test_sig_genes,] + 1) # log transform
#heatmap_data <- t(apply(heatmap_data, 1, function(x){ # Normalize gene expr by dividing by SD
#  x/sd(x)
#}))

#myLower <- 0
#myUpper <- 10

#heatmap_data[heatmap_data < myLower] <- myLower
#heatmap_data[heatmap_data > myUpper] <- myUpper

rownames(heatmap_data) <- lookupGeneName(dat.filtered, rownames(heatmap_data))
heatmap_data <- heatmap_data[order.dendrogram(as.dendrogram(hclust(dist(heatmap_data)))),order.dendrogram(as.dendrogram(hclust(dist(t(heatmap_data)))))]

heatmap_annotation <- pData(dat.filtered)[, c("cluster", "genotype","sex")]

#heatmap_colors <- list(
#"Cell Type" =  c("Acetylcholinergic Neuron" = "#E41A1C", "Neuron" = "#377EB8", "Noradrenergic Neuron" = "#4DAF4A", #"Progenitor/Glia" = "#984EA3", "Transition Cell" = "#FF7F00"),
#"Genotype" = c("het" = "black", "hom" = "gray70"),
#"Sex" = c("female"="black","male"="gray70"),
#"Age" = c("E12.5" = "#7FC97F", "E14.5" = "#BEAED4"))

colnames(heatmap_annotation) <- c("Cluster","Genotype","Sex")

pdf("heatmap_DEG_genotype-sex.pdf", height = 20, width = 20)
pheatmap(mat = heatmap_data,
         scale="none",
         show_rownames = TRUE,
         show_colnames = FALSE,
         drop_levels = FALSE,
         cluster_rows = F,
         cluster_cols = F,
         color = viridis(100),
         annotation_col = heatmap_annotation,
         #annotation_colors = heatmap_colors,
         annotation_names_col = TRUE,
         annotation_legend = TRUE,
         fontsize_row = 5,
         treeheight_col = 0,
         treeheight_row = 0
         )
dev.off()

egENSEMBL <- toTable(org.Mm.egENSEMBL)

genotype_sex_diff_test_entrez <- gsub("\\..*","", genotype_sex_diff_test_sig_genes)
m <- match(genotype_sex_diff_test_entrez, egENSEMBL$ensembl_id)
genotype_sex_diff_test_entrez <- egENSEMBL$gene_id[m]
ego_genotype_sex <- enrichGO(gene = genotype_sex_diff_test_entrez,
                OrgDb =       'org.Mm.eg.db',
                ont           = "BP",
                pAdjustMethod = "BH",
                pvalueCutoff  = 1,
                qvalueCutoff  = 0.05, 
                readable      = TRUE)

write.csv(ego_genotype_sex@result[ego_genotype_sex@result$qvalue < 0.05,], "DEG_all_cells_genotypeXsex_GO.csv")
```

```{r genotype:celltype}
genotypeCellType_diff_test <- differentialGeneTest(dat.filtered[expressed_genes,], fullModelFormulaStr = "~genotype*celltype + batch + total_mRNAs", reducedModelFormulaStr = "~genotype + celltype + batch + total_mRNAs", cores = 1, verbose = T)

genotypeCellType_diff_test_sorted <- genotypeCellType_diff_test[order(genotypeCellType_diff_test$qval),]

write.csv(genotypeCellType_diff_test_sorted[genotypeCellType_diff_test_sorted$qval < 0.05, c("gene_short_name", "gene_id", "pval", "qval")], file = "DEG_all_cells_genotypeXcelltype.csv", quote = F, row.names = F)

genotypeCellType_diff_test_sig_genes<-rownames(genotypeCellType_diff_test_sorted[genotypeCellType_diff_test_sorted$qval < 0.05,])

length(genotypeCellType_diff_test_sig_genes)

genotypeCellType_diff_test_entrez <- gsub("\\..*","", genotypeCellType_diff_test_sig_genes)
m <- match(genotypeCellType_diff_test_entrez, egENSEMBL$ensembl_id)
genotypeCellType_diff_test_entrez <- egENSEMBL$gene_id[m]
ego_genotype_celltype <- enrichGO(gene = genotypeCellType_diff_test_entrez,
                OrgDb =       'org.Mm.eg.db',
                ont           = "BP",
                pAdjustMethod = "BH",
                pvalueCutoff  = 1,
                qvalueCutoff  = 0.05, 
                readable      = TRUE)


ego_genotype_celltype_subset1 <- ego_genotype_celltype@result[ego_genotype_celltype@result$ID %in% 
                                                                c("GO:0007059", "GO:0006260", "GO:0045787", "GO:0044770", "GO:0006281", 
                                                                  "GO:0045786", "GO:1902850", "GO:0007051", "GO:0051383", "GO:0071103", 
                                                                  "GO:0051321", "GO:0033044", "GO:0051656", "GO:0006310", "GO:0034502", 
                                                                  "GO:0000910", "GO:0051302", "GO:0007219", "GO:0042136", "GO:0010718",
                                                                  "GO:0042063"),]

ego_genotype_celltype_subset1$Description <- factor(ego_genotype_celltype_subset1$Description, levels = rev(ego_genotype_celltype_subset1$Description))

pdf("DEG_all_cells_genotypeXcelltype_GO_plot.pdf", height = 3.5, width = 8, useDingbats = F)
ggplot(ego_genotype_celltype_subset1) +
  geom_col(aes(x = Description, y = Count, fill = qvalue)) +
  scale_x_discrete("GO:BP Term") +
  scale_y_continuous("# Genes", expand = c(0,0)) +
  scale_fill_viridis("q-value", limits = c(0, 0.05), breaks = c(0, 0.05), option = "plasma") +
  coord_flip() +
  theme(legend.position = c(0.7, 0.25), 
        legend.background = element_blank(),
        legend.title = element_text(size = 12),
        legend.text = element_text(size = 10),
        axis.text = element_text(size = 10), 
        axis.title = element_text(size = 12),
#        axis.title.y = element_text(margin = margin(r = -10)),
        axis.title.x = element_text(margin = margin(t = -2)))
dev.off()

write.csv(ego_genotype_celltype@result[ego_genotype_celltype@result$qvalue < 0.05,], "DEG_all_cells_genotypeXcelltype_GO.csv")

gene_ids <- str_split(ego_genotype_celltype_subset1$geneID, "\\/")
names(gene_ids) <- str_remove(ego_genotype_celltype_subset1$ID, ":")
tmp <- lapply(gene_ids, function(x){lookupGeneName(dat.filtered, genotypeCellType_diff_test_sig_genes) %in% x})
heatmap_annotation_row <- do.call(cbind, tmp)
heatmap_annotation_row <- as.data.frame(ifelse(heatmap_annotation_row, 1, 0))

S_terms <- c("GO0006260", #DNA replication
         "GO0006281" #DNA repair
)
EarlyMidM_terms <- c("GO1902850", # microtubule cytoskeleton organization involved in mitosis
                  "GO0007051", # spindle organization
                  "GO0051383", # kinetochore organization
                  "GO0007059" # chromosome segregation
                  ) 
LateM_terms <- c("GO0000910", # cytokinesis
                 "GO0051302" # regulation of cell division
                 )
PhaseTransition_terms <- "GO0044770" #cell cycle phase transition



heatmap_annotation_row$S_phase <- apply(heatmap_annotation_row[,S_terms], 1, function(x){
  ifelse(sum(x) > 0, 1, 0)
})

heatmap_annotation_row$EarlyMidM_phase <- apply(heatmap_annotation_row[,EarlyMidM_terms], 1, function(x){
  ifelse(sum(x) > 0, 1, 0)
})

heatmap_annotation_row$LateM_phase <- apply(heatmap_annotation_row[,LateM_terms], 1, function(x){
  ifelse(sum(x) > 0, 1, 0)
})


heatmap_annotation_row <- as.data.frame(apply(heatmap_annotation_row, 2, as.factor))
rownames(heatmap_annotation_row) <- lookupGeneName(dat.filtered, genotypeCellType_diff_test_sig_genes)

heatmap_annotation_row <- heatmap_annotation_row[, c("S_phase", "EarlyMidM_phase", "LateM_phase", "GO0042063", "GO0007219", "GO0042136", "GO0010718")]

heatmap_data <- log10(exprs(dat.filtered)[genotypeCellType_diff_test_sig_genes,] + 1) # log transform
heatmap_data <- t(apply(heatmap_data, 1, function(x){ # Normalize gene expr by dividing by SD
  x/max(x)
}))

rownames(heatmap_data) <- lookupGeneName(dat.filtered,rownames(heatmap_data))
heatmap_data_neurons <- heatmap_data[,rownames(pData(dat.filtered)[pData(dat.filtered)$celltype == "neuron",])]
heatmap_data_neurons <- heatmap_data_neurons[,order.dendrogram(as.dendrogram(hclust(dist(t(heatmap_data_neurons)))))]

heatmap_data_glia <- heatmap_data[,rownames(pData(dat.filtered)[!pData(dat.filtered)$celltype == "neuron",])]
heatmap_data_glia <- heatmap_data_glia[,order.dendrogram(as.dendrogram(hclust(dist(t(heatmap_data_glia)))))]

heatmap_data <- cbind(heatmap_data_neurons, heatmap_data_glia)

heatmap_data <- heatmap_data[order.dendrogram(as.dendrogram(hclust(dist(heatmap_data)))),]

#heatmap_data <- heatmap_data[order.dendrogram(as.dendrogram(hclust(dist(heatmap_data)))),order.dendrogram(as.dendrogram(hclust(dist(t(heatmap_data)))))]

heatmap_annotation_col <- pData(dat.filtered)[,c("celltype","genotype")]

writeLines(paste0("\"", c("GO0007059", "GO0006260", "GO0045787", "GO0044770", "GO0006281", 
               "GO0045786", "GO1902850", "GO0007051", "GO0051383", "GO0071103", 
               "GO0051321", "GO0033044", "GO0051656", "GO0006310", "GO0034502", 
               "GO0000910", "GO0051302", "GO0007219", "GO0042136", "GO0010718",
               "GO0042063"), "\" = c(\"1\" = \"black\", \"0\" = \"grey90\")"), sep = ",\n")

heatmap_colors <- list(
"Cell Type" =  c("neuron" = "#E69F00", "progenitor/glia" = "#56B4E9"),
"Genotype" = c("het" = "#E41A1C", "hom" = "#377EB8"),

"S phase" = c("1" = brewer.pal(7, "Dark2")[1], "0" = "grey90"),
"early-mid M phase" = c("1" = brewer.pal(7, "Dark2")[2], "0" = "grey90"),
"late M phase" = c("1" = brewer.pal(7, "Dark2")[3], "0" = "grey90"),
"gliogenesis" = c("1" = brewer.pal(7, "Dark2")[4], "0" = "grey90"),
"Notch signaling" = c("1" = brewer.pal(7, "Dark2")[5], "0" = "grey90"),
"neurotransmitter biosynthetic process" = c("1" = brewer.pal(7, "Dark2")[6], "0" = "grey90"),
"epithelial to mesenchymal transition" = c("1" = brewer.pal(7, "Dark2")[7], "0" = "grey90")
)

colnames(heatmap_annotation_col) <- c("Cell Type","Genotype")
colnames(heatmap_annotation_row) <- c("S phase", "early-mid M phase", "late M phase", "gliogenesis", "Notch signaling", "neurotransmitter biosynthetic process", "epithelial to mesenchymal transition")

pdf("heatmap_DEG_genotype-celltype.pdf", height = 20, width = 20)
pheatmap(mat = heatmap_data,
         scale="none",
         show_rownames = FALSE,
         show_colnames = FALSE,
         drop_levels = FALSE,
         cluster_rows = F,
         cluster_cols = F,
         color = viridis(100),
         annotation_col = heatmap_annotation_col,
         annotation_row = heatmap_annotation_row,
         annotation_colors = heatmap_colors,
         annotation_names_col = TRUE,
         annotation_legend = TRUE
         )
dev.off()
```

```{r neuron_bifurcation_diff_test}
neuron_bifuracation_difftest <- differentialGeneTest(dat.filtered[expressed_genes,pData(dat.filtered)$cluster %in% 7:8], fullModelFormulaStr = "~cluster + batch + total_mRNAs", reducedModelFormulaStr = "~batch + total_mRNAs", cores = 1, verbose = T)

neuron_bifuracation_difftest_sorted <- neuron_bifuracation_difftest[order(neuron_bifuracation_difftest$qval),]

write.csv(neuron_bifuracation_difftest_sorted[neuron_bifuracation_difftest_sorted$qval < 0.05, c("gene_short_name", "gene_id", "pval", "qval")], file = "DEG_clusters7&8_cluster.csv", quote = F, row.names = F)

neuron_bifuracation_difftest_sig_genes <- rownames(neuron_bifuracation_difftest_sorted[neuron_bifuracation_difftest_sorted$qval < 0.05,])

bifurcation_gene_names <- lookupGeneName(dat.filtered, neuron_bifuracation_difftest_sig_genes)

#tmp <- t(scale(t(exprs(dat.filtered[neuron_bifuracation_difftest_sorted_sig_genes, pData(dat.filtered)$cluster %in% 7:8])), scale = T, center = T))
#heatmap_data <- log10(tmp - min(tmp) + 1)

heatmap_data <- log10(exprs(dat.filtered)[neuron_bifuracation_difftest_sig_genes, pData(dat.filtered)$cluster %in% 7:8] + 1) # log transform
heatmap_data <- t(apply(heatmap_data, 1, function(x){ # Normalize gene expr by dividing by SD
  x/max(x)
}))

rownames(heatmap_data) <- lookupGeneName(dat.filtered, rownames(heatmap_data))

heatmap_data <- heatmap_data[order.dendrogram(as.dendrogram(hclust(dist(heatmap_data)))),order.dendrogram(as.dendrogram(hclust(dist(t(heatmap_data)))))]

#myLower <- 0.1
#myUpper <- 0.9

#heatmap_data[heatmap_data < myLower] <- myLower
#heatmap_data[heatmap_data > myUpper] <- myUpper

heatmap_annotation <- pData(dat.filtered)[pData(dat.filtered)$cluster %in% 7:8, c("cluster", "genotype")]
heatmap_annotation$cluster <- droplevels(heatmap_annotation$cluster)

heatmap_colors <- list(
  "Cluster" = c("7" = brewer.pal(8, "Set2")[7],
                "8" = brewer.pal(8, "Set2")[8]),
  "Genotype" = c("het" = brewer.pal(3, "Set1")[1], 
                 "hom" = brewer.pal(3, "Set1")[2]))

colnames(heatmap_annotation)<-c("Cluster","Genotype")

pdf("heatmap_DEG_cluster7vs8_clustered.pdf", height = 10, width = 14)
pheatmap(mat = heatmap_data,
         scale="none",
         show_rownames = TRUE,
         show_colnames = FALSE,
         drop_levels = FALSE,
         cluster_rows = T,
         cluster_cols = T,
         color = viridis(100),
         annotation_col = heatmap_annotation,
         annotation_colors = heatmap_colors,
         annotation_names_col = TRUE,
         annotation_legend = TRUE,
         fontsize_row = 5,
         treeheight_col = 0,
         treeheight_row = 20,
         cutree_rows = 10
         )
dev.off()

egENSEMBL <- toTable(org.Mm.egENSEMBL)

bifurcation_entrez <- gsub("\\..*","",neuron_bifuracation_difftest_sig_genes)
m<-match(bifurcation_entrez,egENSEMBL$ensembl_id)
bifurcation_entrez <- egENSEMBL$gene_id[m]
ego_bifurcation <- enrichGO(gene = bifurcation_entrez,
                OrgDb =       'org.Mm.eg.db',
                ont           = "BP",
                pAdjustMethod = "BH",
                pvalueCutoff  = 1,
                qvalueCutoff  = 0.05, 
                readable      = TRUE)

write.csv(ego_bifurcation@result[ego_bifurcation@result$qvalue < 0.05,], "DEG_clusters7&8_cluster_GO.csv")
```

```{r Ret_independent_genotype_diff}
Ret_independent_difftest <- differentialGeneTest(dat.filtered[expressed_genes,pData(dat.filtered)$cluster == 8], fullModelFormulaStr = "~genotype + batch + total_mRNAs", reducedModelFormulaStr = "~batch + total_mRNAs", cores = 1, verbose = T)

Ret_independent_difftest_sorted <- Ret_independent_difftest[order(Ret_independent_difftest$qval),]

write.csv(Ret_independent_difftest_sorted[Ret_independent_difftest_sorted$qval < 0.05, c("gene_short_name", "gene_id", "pval", "qval")], file = "DEG_cluster8_genotype.csv", quote = F, row.names = F)

Ret_independent_difftest_sig_genes <- rownames(Ret_independent_difftest_sorted[Ret_independent_difftest_sorted$qval < 0.05,])

het_mean <- apply(exprs(dat.filtered[Ret_independent_difftest_sig_genes, pData(dat.filtered)$cluster == 8 & pData(dat.filtered)$genotype == "het"]), 1, mean)

hom_mean <- apply(exprs(dat.filtered[Ret_independent_difftest_sig_genes, pData(dat.filtered)$cluster == 8 & pData(dat.filtered)$genotype == "hom"]), 1, mean)


lookupGeneName(dat.filtered, names(which(het_mean/hom_mean < 1)))
# "Tac1" "Pcp4" "Ddc" "Th" "A730017C20Rik" "Rgs4" "Agtr2" "Spock1"  higher in null

lookupGeneName(dat.filtered, names(which(het_mean/hom_mean > 1)))
#"Gm27747" "Akr1e1"  "Gm27553" "Hoxb5" higher in het

Ret_independent_gene_names <- lookupGeneName(dat.filtered, Ret_independent_difftest_sig_genes)


#tmp <- t(scale(t(exprs(dat.filtered[Ret_independent_difftest_sorted_sig_genes, pData(dat.filtered)$cluster %in% c(4, 6)])), scale = T, center = T))
#heatmap_data <- log10(tmp - min(tmp) + 1)

heatmap_data <- log10(exprs(dat.filtered)[Ret_independent_difftest_sig_genes, pData(dat.filtered)$cluster == 8] + 1) # log transform
heatmap_data <- t(apply(heatmap_data, 1, function(x){ # Normalize gene expr by dividing by max
  x/max(x)
}))

rownames(heatmap_data) <- lookupGeneName(dat.filtered, rownames(heatmap_data))

heatmap_data <- heatmap_data[order.dendrogram(as.dendrogram(hclust(dist(heatmap_data)))),order.dendrogram(as.dendrogram(hclust(dist(t(heatmap_data)))))]

#myLower <- 0.1
#myUpper <- 0.9

#heatmap_data[heatmap_data < myLower] <- myLower
#heatmap_data[heatmap_data > myUpper] <- myUpper

heatmap_annotation <- pData(dat.filtered)[pData(dat.filtered)$cluster == 8, c("genotype"), drop = F]


heatmap_colors <- list(
  "Genotype" = c("het" = brewer.pal(3, "Set1")[1], 
                 "hom" = brewer.pal(3, "Set1")[2]))

colnames(heatmap_annotation)<-c("Genotype")

pdf("heatmap_DEG_genotype_cluster8.pdf", height = 3, width = 8)
pheatmap(mat = heatmap_data,
         scale="none",
         show_rownames = TRUE,
         show_colnames = FALSE,
         drop_levels = FALSE,
         cluster_rows = F,
         cluster_cols = F,
         color = viridis(100),
         annotation_col = heatmap_annotation,
         annotation_colors = heatmap_colors,
         annotation_names_col = TRUE,
         annotation_legend = TRUE,
         fontsize_row = 7,
         treeheight_col = 0,
         treeheight_row = 0
         )
dev.off()

egENSEMBL <- toTable(org.Mm.egENSEMBL)

Ret_independent_entrez <- gsub("\\..*","",Ret_independent_difftest_sig_genes)
m<-match(Ret_independent_entrez,egENSEMBL$ensembl_id)
Ret_independent_entrez <- egENSEMBL$gene_id[m]
ego_Ret_independent <- enrichGO(gene = Ret_independent_entrez,
                OrgDb =       'org.Mm.eg.db',
                ont           = "BP",
                pAdjustMethod = "BH",
                pvalueCutoff  = 1,
                qvalueCutoff  = 0.05, 
                readable      = TRUE)

write.csv(ego_Ret_independent@result[ego_Ret_independent@result$qvalue < 0.05,], "DEG_cluster8_genotype_GO.csv")
```

```{r genotype_glia}
glia_genotype_diff_test <- differentialGeneTest(dat.filtered[expressed_genes, pData(dat.filtered)$celltype == "progenitor/glia" & !pData(dat.filtered)$cluster == 9], fullModelFormulaStr = "~genotype + batch + total_mRNAs", reducedModelFormulaStr = "~batch + total_mRNAs", verbose = T)

glia_genotype_diff_test_sorted <- glia_genotype_diff_test[order(glia_genotype_diff_test$qval),]

write.csv(glia_genotype_diff_test_sorted[glia_genotype_diff_test_sorted$qval < 0.05, c("gene_short_name", "gene_id", "pval", "qval")], file = "DEG_glia!cluster9_genotype.csv", quote = F, row.names = F)

glia_genotype_diff_test_sig_genes <- rownames(glia_genotype_diff_test_sorted[glia_genotype_diff_test_sorted$qval < 0.05,])

heatmap_data <- log10(exprs(dat.filtered)[glia_genotype_diff_test_sig_genes, pData(dat.filtered)$celltype == "progenitor/glia" & !pData(dat.filtered)$cluster == 9] + 1) # log transform
heatmap_data <- t(apply(heatmap_data, 1, function(x){ # Normalize gene expr by dividing by max
  x/max(x)
}))

rownames(heatmap_data) <- lookupGeneName(dat.filtered, rownames(heatmap_data))

heatmap_data <- heatmap_data[order.dendrogram(as.dendrogram(hclust(dist(heatmap_data)))),order.dendrogram(as.dendrogram(hclust(dist(t(heatmap_data)))))]

#myLower <- 0.1
#myUpper <- 0.9

#heatmap_data[heatmap_data < myLower] <- myLower
#heatmap_data[heatmap_data > myUpper] <- myUpper

heatmap_annotation <- pData(dat.filtered)[pData(dat.filtered)$celltype == "progenitor/glia" & !pData(dat.filtered)$cluster == 9, c("genotype", "cluster")]


heatmap_colors <- list(
  "Genotype" = c("het" = brewer.pal(3, "Set1")[1], 
                 "hom" = brewer.pal(3, "Set1")[2]),
  "Cluster" = c("1" = brewer.pal(8, "Set2")[1], 
                "2" = brewer.pal(8, "Set2")[2], 
                "3" = brewer.pal(8, "Set2")[3], 
                "4" = brewer.pal(8, "Set2")[4], 
                "5" = brewer.pal(8, "Set2")[5], 
                "6" = brewer.pal(8, "Set2")[6])
  )

colnames(heatmap_annotation) <- c("Genotype", "Cluster")

pdf("DEG_glia!cluster9_genotype_heatmap.pdf", height = 24, width = 20)
pheatmap(mat = heatmap_data,
         scale="none",
         show_rownames = TRUE,
         show_colnames = FALSE,
         drop_levels = FALSE,
         cluster_rows = T,
         cluster_cols = T,
         color = viridis(100),
         annotation_col = heatmap_annotation,
         annotation_colors = heatmap_colors,
         annotation_names_col = TRUE,
         annotation_legend = TRUE,
         fontsize_row = 2,
         treeheight_col = 60,
         treeheight_row = 0
         )
dev.off()

egENSEMBL <- toTable(org.Mm.egENSEMBL)

glia_genotype_entrez <- gsub("\\..*","",glia_genotype_diff_test_sig_genes)
m<-match(glia_genotype_entrez,egENSEMBL$ensembl_id)
glia_genotype_entrez <- egENSEMBL$gene_id[m]
ego_glia_genotype <- enrichGO(gene = glia_genotype_entrez,
                OrgDb =       'org.Mm.eg.db',
                ont           = "BP",
                pAdjustMethod = "BH",
                pvalueCutoff  = 1,
                qvalueCutoff  = 0.05, 
                readable      = TRUE)

ego_glia_genotype_significant <- ego_glia_genotype@result[ego_glia_genotype@result$qvalue < 0.05,]

write.csv(ego_glia_genotype@result[ego_glia_genotype@result$qvalue < 0.05,], "DEG_glia!cluster9_genotype_GO.csv")

glia_terms <- c("GO:0042063", "GO:0045165", "GO:0014033", "GO:0061351", "GO:0045787", 
                "GO:0042552", "GO:0007088", "GO:0045786", "GO:0000132", "GO:0051302", 
                "GO:0000082")

ego_glia_subset <- ego_glia_genotype@result[glia_terms,]

ego_glia_subset$Description <- factor(ego_glia_subset$Description, levels = rev(ego_glia_subset$Description))

pdf("F4B-DEG_glia!cluster9_genotype_GO_plot.pdf", height = 2.5, width = 6, useDingbats = F)
ggplot(ego_glia_subset) +
  geom_col(aes(x = Description, y = Count, fill = qvalue)) +
  scale_x_discrete("GO:BP Term") +
  scale_y_continuous("# Genes", expand = c(0,0)) +
  scale_fill_viridis("q-value", limits = c(0, 0.05), breaks = c(0, 0.05), option = "plasma") +
  coord_flip() +
  theme(legend.position = c(0.77, 0.35), 
        legend.background = element_blank(),
        legend.title = element_text(size = 12),
        legend.text = element_text(size = 10),
        axis.text = element_text(size = 10), 
        axis.title = element_text(size = 12),
        axis.title.y = element_text(margin = margin(r = -10), hjust = 0.6),
        axis.title.x = element_text(margin = margin(t = -2)))
dev.off()
```
